refactor: Split engine builder modules
